AppVeyor CI integration
We add basic AppVeyor CI integration for building the extension and
running the tests.  We add jobs for all twelve variations currently
built on PECL, plus a single job for each PHP 7.0 and 7.1 for the most
typical variation on Windows (x64, NTS).  We do not add jobs for PHP 5,
mostly because AppVeyor does not support Visual Studio 2011 which would
be required for PHP 5.5 and 5.6, and also because we would need to
fiddle with the PHP-SDK v1.  Furthermore, 14 jobs already seem to be
a lot.